refactor: rename root Loading components dir
