+++ renderer/components/Loading/LoadingApp.js
