+++ renderer/components/Loading/LoadingBolt.js
