+++ renderer/components/Loading/LoadingLaunchpad.js
