+++ renderer/components/Loading/index.js
