+++ renderer/containers/Loading/Loading.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import Loading from 'components/LoadingBolt'
+import Loading from 'components/Loading'
 import { walletSelectors } from 'reducers/wallet'
 
 const LoadingContainer = ({ pathname, hasClose, isLoading, hasWallets, isWalletOpen, onClose }) => {
