+++ test/unit/components/LoadingBolt.spec.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { ThemeProvider } from 'styled-components'
 import { RawIntlProvider } from 'react-intl'
 import { mount } from 'enzyme'
-import LoadingBolt from 'components/LoadingBolt'
+import LoadingBolt from 'components/Loading/LoadingBolt'
 import CloudLightning from 'components/Icon/CloudLightning'
 import { dark } from 'themes'
 import intl from '../__helpers__/intl'
